blacklist: port to generic layer.
