+++ Library/Homebrew/extend/os/blacklist.rb
@@ -1,0 +1,5 @@
+require "blacklist"
+
+if OS.mac?
+  require "extend/os/mac/blacklist"
+end
