+++ Library/Homebrew/test/test_os_mac_blacklist.rb
@@ -1,0 +1,12 @@
+require "testing_env"
+require "blacklist"
+
+class BlacklistTests < Homebrew::TestCase
+  def assert_blacklisted(s)
+    assert blacklisted?(s), "'#{s}' should be blacklisted"
+  end
+
+  def test_xcode
+    %w[xcode Xcode].each { |s| assert_blacklisted s }
+  end
+end
